Hunts and Notifiers API Standardization and Documentation (#1571)
* start breaking out hunt endpoint functions

* move more stuff to apiHunts

* move more stuff over to apiHunts

create apiNotifiers too

* hunt endpoint standardization and documentation wrap up

* move all notifier stuff into apinotifier file

document it
standardize endpoints

* use singular noun for create notifier endpoint

* add more info to the hunts help area

also update some docs to reference the corresponding typdefs

* fix hunt and notifier tests

change some endpoints to new ones

* use old api endpoint for remote hunts for now

* wrap console logs in debug check in apihunts

* show console logs for errors
