+++ tests/api-notifiers.t
@@ -1,4 +1,4 @@
-use Test::More tests => 33;
+use Test::More tests => 31;
 use Cwd;
 use URI::Escape;
 use MolochTest;
@@ -11,75 +11,71 @@ my $token = getTokenCookie();
 my $notAdminToken = getTokenCookie('notadmin');
 
 # notifier types
-  my $notifierTypes = viewerGetToken("/notifierTypes", $token);
+  my $notifierTypes = viewerGetToken("/api/notifiertypes", $token);
   ok(exists $notifierTypes->{email}, "email notifier exists");
   ok(exists $notifierTypes->{slack}, "slack notifier exists");
   ok(exists $notifierTypes->{twilio}, "twilio notifier exists");
 
 # empty notifiers
-  my $notifiers = viewerGetToken("/notifiers", $token);
+  my $notifiers = viewerGetToken("/api/notifiers", $token);
   ok(!keys %{$notifiers}, "Empty notifiers");
 
 # create notifier required items
-  my $json = viewerPostToken("/notifiers", '{}', $token);
-  is($json->{text}, "Missing notifier", "notifier object required");
-  $json = viewerPostToken("/notifiers", '{"notifier":{}}', $token);
+  my $json = viewerPostToken("/api/notifier", '{}', $token);
   is($json->{text}, "Missing a unique notifier name", "notifier name required");
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1"}}', $token);
+  $json = viewerPostToken("/notifiers", '{"name":"test1"}', $token);
   is($json->{text}, "Missing notifier type", "notifier type required");
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1","type":"slack"}}', $token);
+  $json = viewerPostToken("/api/notifier", '{"name":"test1","type":"slack"}', $token);
   is($json->{text}, "Missing notifier fields", "notifier fields required");
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1","type":"slack","fields":"badfields"}}', $token);
+  $json = viewerPostToken("/notifiers", '{"name":"test1","type":"slack","fields":"badfields"}', $token);
   is($json->{text}, "Notifier fields must be an array", "notifier fields must be an array");
 
 # create notifier requires token and admin access
   $json = viewerPost("/notifiers", '{}');
   is($json->{text}, "Missing token", "create notifier requires token");
   $json = viewerPostToken("/notifiers?molochRegressionUser=notadmin", '{}', $notAdminToken);
-  is($json->{text}, "Need admin privelages to create a notifier", "create notifier requires admin");
+  is($json->{text}, "You do not have permission to access this resource", "create notifier requires admin");
 
 # create notifier needs valid notifier type
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1","type":"unknown","fields":[]}}', $token);
+  $json = viewerPostToken("/api/notifier", '{"name":"test1","type":"unknown","fields":[]}', $token);
   is($json->{text}, "Unknown notifier type", "invalid notifier type");
 
 # create notifier
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}}', $token);
+  $json = viewerPostToken("/notifiers", '{"name":"test1","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}', $token);
   ok($json->{success}, "notifier create success");
 
 # create notifier requires unique notifier name
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test1","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}}', $token);
+  $json = viewerPostToken("/notifiers", '{"name":"test1","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}', $token);
   is($json->{text}, "Notifier already exists", "notifier must have a unique name");
 
 # create notifier sanitizes notifier name
-  $json = viewerPostToken("/notifiers", '{"notifier":{"name":"test2`~!@#$%^&*+[]{}(),.<>?","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}}', $token);
+  $json = viewerPostToken("/api/notifier", '{"name":"test2`~!@#$%^&*+[]{}(),.<>?","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1url"}}]}', $token);
   is($json->{name}, "test2", "notifier name sanitization");
 
 # update notifier requires admin access
-  $json = viewerPutToken("/notifiers/test1?molochRegressionUser=notadmin", '{}', $notAdminToken);
-  is($json->{text}, "Need admin privelages to update a notifier", "update notifier requires admin");
+  $json = viewerPutToken("/api/notifier/test1?molochRegressionUser=notadmin", '{}', $notAdminToken);
+  is($json->{text}, "You do not have permission to access this resource", "update notifier requires admin");
 
 # update notifier needs valid name
   $json = viewerPutToken("/notifiers/badname", '{}', $token);
   is($json->{text}, "Cannot find notifer to udpate", "update notifier needs valid name");
 
 # update notifier required fields
-  $json = viewerPutToken("/notifiers/test1", '{}', $token);
-  is($json->{text}, "Missing notifier", "notifier object required");
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{}}', $token);
+  $json = viewerPutToken("/api/notifier/test1", '{}', $token);
   is($json->{text}, "Missing a unique notifier name", "notifier name required");
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{"name":"test1a"}}', $token);
+  $json = viewerPutToken("/notifiers/test1", '{"name":"test1a"}', $token);
   is($json->{text}, "Missing notifier type", "notifier type required");
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{"name":"test1a","type":"slack"}}', $token);
+  $json = viewerPutToken("/api/notifier/test1", '{"name":"test1a","type":"slack"}', $token);
   is($json->{text}, "Missing notifier fields", "notifier fields required");
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{"name":"test1a","type":"slack","fields":"badfields"}}', $token);
+  $json = viewerPutToken("/notifiers/test1", '{"name":"test1a","type":"slack","fields":"badfields"}', $token);
   is($json->{text}, "Notifier fields must be an array", "notifier fields must be an array");
 
 # update notifier needs valid notifier type
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{"name":"test1a","type":"unknown","fields":[]}}', $token);
+  $json = viewerPutToken("/notifiers/test1", '{"name":"test1a","type":"unknown","fields":[]}', $token);
   is($json->{text}, "Unknown notifier type", "invalid notifier type");
 
 # update notifier
-  $json = viewerPutToken("/notifiers/test1", '{"notifier":{"name":"test1a","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1aurl"}}]}}', $token);
+  $json = viewerPutToken("/notifiers/test1", '{"name":"test1a","type":"slack","fields":[{"slackWebhookUrl":{"value":"test1aurl"}}]}', $token);
   ok($json->{success}, "notifier update success");
   is($json->{name}, "test1a", "notifier name update");
   $notifiers = viewerGetToken("/notifiers", $token);
@@ -92,7 +88,7 @@ my $notAdminToken = getTokenCookie('notadmin');
   ok(!exists $notifiers->{test1a}->{fields}, "fields shouldn't exist for non admin");
 
 # cleanup
-  $json = viewerDeleteToken("/notifiers/test1a", $token);
+  $json = viewerDeleteToken("/api/notifier/test1a", $token);
   ok($json->{success}, "notifier delete success");
   $json = viewerDeleteToken("/notifiers/test2", $token);
   ok($json->{success}, "notifier delete success");
